Introduce Art-10 bot
Bug: 144966342

+++ scripts/update-host-art.sh
@@ -109,7 +109,10 @@ cp $ANDROID_HOST_BUILD/bin/dalvikvm32 $DEST/bin
 # dalvikvm32 or dalvikvm64).
 cp $ANDROID_HOST_BUILD/bin/dalvikvm $DEST/bin
 cp $ANDROID_HOST_BUILD/bin/dex2oat $DEST/bin
-cp $ANDROID_HOST_BUILD/bin/patchoat $DEST/bin
+if [ -e $ANDROID_HOST_BUILD/bin/patchoat ]
+  # File patchoat does not exist on Q anymore.
+  then cp $ANDROID_HOST_BUILD/bin/patchoat $DEST/bin
+fi
 
 # Required framework files.
 mkdir -p $DEST/framework
@@ -127,8 +130,13 @@ mkdir -p $DEST/product/$ANDROID_PRODUCT/system/framework
 cp -r $ANDROID_TARGET_BUILD/product/$ANDROID_PRODUCT/system/framework/* $DEST/product/$ANDROID_PRODUCT/system/framework
 
 # Required auxillary files.
-mkdir -p $DEST/usr/icu
-cp -r $ANDROID_HOST_BUILD/usr/icu/* $DEST/usr/icu
+if [ -e $ANDROID_HOST_BUILD/usr/icu ]; then
+  mkdir -p $DEST/usr/icu
+  cp -r $ANDROID_HOST_BUILD/usr/icu/* $DEST/usr/icu
+else
+  mkdir -p $DEST/com.android.runtime/etc/icu/
+  cp -r $ANDROID_HOST_BUILD/com.android.runtime/etc/icu/* $DEST/com.android.runtime/etc/icu/
+fi
 
 # Update links for vdex files for Android P and later.
 if [ -f $DEST/product/$ANDROID_PRODUCT/system/framework/boot.vdex ]; then
